Added loading suspense for leaderboard page
