+++ apps/web/src/app/(navigation)/leaderboards/Leaderboard.tsx
@@ -1,0 +1,53 @@
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from "@/components/ui/table";
+import { cache } from "@/lib/cache";
+import { delay } from "@/lib/delay";
+import { prisma } from "@/lib/prisma";
+
+const getAllPlayersByGains = cache(async () => {
+  return await prisma.player.findMany({ orderBy: { gains: "desc" } });
+}, ["/leaderboards"]);
+
+export default async function Leaderboard() {
+  const allPlayers = await getAllPlayersByGains();
+  return (
+    <>
+      <h1>Leaderboards</h1>
+
+      <div>
+        <Table>
+          <TableHeader>
+            <TableRow>
+              <TableHead>Place</TableHead>
+              <TableHead>Player</TableHead>
+              <TableHead>Gains</TableHead>
+            </TableRow>
+          </TableHeader>
+          <TableBody>
+            {allPlayers.map((player, index) => {
+              return (
+                <TableRow key={player.id}>
+                  <TableCell>{index + 1}</TableCell>
+                  <TableCell>
+                    {index === 0 ? (
+                      <span>{player.name} ( Aura Farmer 👑 )</span>
+                    ) : (
+                      player.name
+                    )}
+                  </TableCell>
+                  <TableCell>{+player.gains}</TableCell>
+                </TableRow>
+              );
+            })}
+          </TableBody>
+        </Table>
+      </div>
+    </>
+  );
+}
