+++ apps/web/src/app/(navigation)/leaderboards/LeaderboardSuspense.tsx
@@ -1,0 +1,49 @@
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from "@/components/ui/table";
+
+export default function LeaderboardSuspense() {
+  const skeletonRows = Array.from({ length: 5 });
+
+  return (
+    <div>
+      <Table>
+        <TableHeader>
+          <TableRow>
+            <TableHead>Place</TableHead>
+            <TableHead>Player</TableHead>
+            <TableHead>Gains</TableHead>
+          </TableRow>
+        </TableHeader>
+        <TableBody>
+          {skeletonRows.map((_, index) => (
+            <TableRow key={index}>
+              <TableCell>
+                <Skeleton className="h-4 w-6" />
+              </TableCell>
+              <TableCell>
+                <Skeleton className="h-4 w-40" />
+              </TableCell>
+              <TableCell>
+                <Skeleton className="h-4 w-12" />
+              </TableCell>
+            </TableRow>
+          ))}
+        </TableBody>
+      </Table>
+    </div>
+  );
+}
+
+function Skeleton({ className }: { className?: string }) {
+  return (
+    <div
+      className={`animate-pulse bg-muted rounded-md ${className || "h-4 w-full"}`}
+    />
+  );
+}
